♻️ Code refactor+fix
